+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.2 Release 
+
+* Initial Release 
+* Added option to show Logo in drawer 
+* Added title in drawer 
+* Added subtitle in drawe 
+* Added Sample Project 
+
 ## 1.0.1 First Test Release
 
 * First Test Release